Add HTML parser
The parser converts HTML to VDOM models.
