+++ src/py/idom/widgets/__init__.py
@@ -1,6 +1,7 @@
-from .common import Var, node, Image, hotswap
+from .common import node, hotswap
 from .display import display
 from .inputs import Input
+from .images import Image
 from . import html
 
-__all__ = ["display", "Var", "node", "Image", "hotswap", "html", "Input"]
+__all__ = ["display", "node", "Image", "hotswap", "html", "Input"]
